requirements.txt
celery==5.4.0
django-openid-auth==0.17
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-OpenID==1.3.1
Flask-Pydantic==0.12.0
Flask-SQLAlchemy==3.1.1
gunicorn[eventlet]==23.0.0
google-api-python-client==2.78.0
multiprocess==0.70.17
psycopg2-binary==2.9.10
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
redis[hiredis]==5.2.1